About this first iteration
This update is using the perlin noise to have some movement from the circle base X and Y,, and at the same time This noise value is stored in the variable noiseFactorAngle, This noise value is used to calculate the new angle newAngle. This is done by taking the current angle this.angle and adding the noiseFactorAngle multiplied by π (pi). The use of π here means that the noise value is mapped to a full circle (0 to 360 degrees), allowing the shape to rotate in a full circle. After calculating the new angle newAngle, this angle value is passed to the circleNeon function as the last parameter. In the circleNeon function, this angle value should be used to determine how to draw or rotate the graphics. Although the specific implementation of the circleNeon function is not provided in the code snippet, it is reasonable to infer that it uses this angle to rotate the drawn graphics.