About this second iteration
I added some rains and shining effect  to stimulate the lightning effect. My inspiration was to make this tree alive, and the movement is like a live tree, which could be seen more horrible just like a scary character which often show up at a rainy night, there will be more explanation in the readme file, but the code is really simple just used some random function yo identify it.